test fixes

+++ services/repository-managers/src/test/java/org/sagebionetworks/repo/manager/file/FileHandleManagerImplAutowireTest.java
@@ -4,6 +4,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 import static org.mockito.Mockito.when;
 
 import java.io.IOException;
@@ -17,6 +18,7 @@ import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Random;
+import java.util.Set;
 import java.util.UUID;
 
 import org.apache.commons.fileupload.FileItemIterator;
@@ -33,16 +35,23 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mockito;
 import org.sagebionetworks.StackConfiguration;
+import org.sagebionetworks.StackConstants;
 import org.sagebionetworks.downloadtools.FileUtils;
 import org.sagebionetworks.repo.manager.EntityManager;
 import org.sagebionetworks.repo.manager.ProjectSettingsManager;
 import org.sagebionetworks.repo.manager.S3TokenManagerImpl;
 import org.sagebionetworks.repo.manager.UserManager;
 import org.sagebionetworks.repo.manager.file.transfer.TransferUtils;
+import org.sagebionetworks.repo.model.ACCESS_TYPE;
+import org.sagebionetworks.repo.model.AccessControlList;
+import org.sagebionetworks.repo.model.AccessControlListDAO;
 import org.sagebionetworks.repo.model.AuthorizationConstants.BOOTSTRAP_PRINCIPAL;
 import org.sagebionetworks.repo.model.DatastoreException;
 import org.sagebionetworks.repo.model.Folder;
+import org.sagebionetworks.repo.model.ObjectType;
 import org.sagebionetworks.repo.model.Project;
+import org.sagebionetworks.repo.model.ResourceAccess;
+import org.sagebionetworks.repo.model.UnauthorizedException;
 import org.sagebionetworks.repo.model.UserInfo;
 import org.sagebionetworks.repo.model.attachment.AttachmentData;
 import org.sagebionetworks.repo.model.auth.NewUser;
@@ -60,26 +69,30 @@ import org.sagebionetworks.repo.model.file.UploadDaemonStatus;
 import org.sagebionetworks.repo.model.file.UploadDestination;
 import org.sagebionetworks.repo.model.file.UploadDestinationLocation;
 import org.sagebionetworks.repo.model.file.UploadType;
+import org.sagebionetworks.repo.model.project.ExternalS3UploadDestinationLocationSetting;
 import org.sagebionetworks.repo.model.project.ExternalUploadDestinationLocationSetting;
-import org.sagebionetworks.repo.model.project.ExternalUploadDestinationSetting;
 import org.sagebionetworks.repo.model.project.ProjectSettingsType;
 import org.sagebionetworks.repo.model.project.UploadDestinationListSetting;
-import org.sagebionetworks.repo.model.project.UploadDestinationSetting;
 import org.sagebionetworks.repo.web.NotFoundException;
 import org.sagebionetworks.repo.web.ServiceUnavailableException;
 import org.sagebionetworks.utils.DefaultHttpClientSingleton;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import com.amazonaws.services.s3.AmazonS3Client;
+import com.amazonaws.services.s3.internal.Constants;
+import com.amazonaws.services.s3.model.Bucket;
 import com.amazonaws.services.s3.model.BucketCrossOriginConfiguration;
 import com.amazonaws.services.s3.model.CORSRule;
 import com.amazonaws.services.s3.model.CORSRule.AllowedMethods;
+import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.amazonaws.services.s3.model.S3Object;
 import com.amazonaws.util.BinaryUtils;
 import com.amazonaws.util.StringInputStream;
 import com.google.common.collect.Lists;
+import com.google.common.collect.Sets;
 
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(locations = { "classpath:test-context.xml" })
@@ -107,8 +120,13 @@ public class FileHandleManagerImplAutowireTest {
 
 	@Autowired
 	private ProjectSettingsManager projectSettingsManager;
+
+	@Autowired
+	private AccessControlListDAO accessControlListDAO;
 	
 	private UserInfo userInfo;
+	private UserInfo userInfo2;
+	private String username;
 	
 	/**
 	 * This is the metadata about the files we uploaded.
@@ -125,16 +143,27 @@ public class FileHandleManagerImplAutowireTest {
 	@Before
 	public void before() throws Exception{
 		NewUser user = new NewUser();
-		user.setEmail(UUID.randomUUID().toString() + "@test.com");
-		user.setUserName(UUID.randomUUID().toString());
+		username = UUID.randomUUID().toString();
+		user.setEmail(username + "@test.com");
+		user.setUserName(username);
 		userInfo = userManager.getUserInfo(userManager.createUser(user));
 		userInfo.getGroups().add(BOOTSTRAP_PRINCIPAL.CERTIFIED_USERS.getPrincipalId());
+
+		user = new NewUser();
+		String username2 = UUID.randomUUID().toString();
+		user.setEmail(username2 + "@test.com");
+		user.setUserName(username2);
+		userInfo2 = userManager.getUserInfo(userManager.createUser(user));
+		userInfo2.getGroups().add(BOOTSTRAP_PRINCIPAL.CERTIFIED_USERS.getPrincipalId());
 		
 		Project project = new Project();
 		projectName = "project" + new Random().nextInt();
 		project.setName(projectName);
 		projectId = entityManager.createEntity(userInfo, project, null);
 		entitiesToDelete.add(projectId);
+
+		addAcl(projectId, userInfo2.getId());
+
 		Folder child = new Folder();
 		child.setName("child");
 		child.setParentId(projectId);
@@ -179,8 +208,10 @@ public class FileHandleManagerImplAutowireTest {
 			for(S3FileHandle meta: toDelete){
 				// delete the file from S3.
 				s3Client.deleteObject(meta.getBucketName(), meta.getKey());
-				// We also need to delete the data from the database
-				fileHandleDao.delete(meta.getId());
+				if (meta.getId() != null) {
+					// We also need to delete the data from the database
+					fileHandleDao.delete(meta.getId());
+				}
 			}
 		}
 		for (String id : Lists.reverse(entitiesToDelete)) {
@@ -189,6 +220,7 @@ public class FileHandleManagerImplAutowireTest {
 		
 		UserInfo adminUserInfo = userManager.getUserInfo(BOOTSTRAP_PRINCIPAL.THE_ADMIN_USER.getPrincipalId());
 		userManager.deletePrincipal(adminUserInfo, Long.parseLong(userInfo.getId().toString()));
+		userManager.deletePrincipal(adminUserInfo, Long.parseLong(userInfo2.getId().toString()));
 	}
 	
 
@@ -408,7 +440,9 @@ public class FileHandleManagerImplAutowireTest {
 
 		projectSettingsManager.createProjectSetting(userInfo, uploadDestinationListSetting);
 
+		@SuppressWarnings("deprecation")
 		List<UploadDestination> uploadDestinations = fileUploadManager.getUploadDestinations(userInfo, uploadFolder);
+
 		assertEquals(1, uploadDestinations.size());
 		assertEquals(ExternalUploadDestination.class, uploadDestinations.get(0).getClass());
 		ExternalUploadDestination externalUploadDestination = (ExternalUploadDestination) uploadDestinations.get(0);
@@ -451,6 +485,75 @@ public class FileHandleManagerImplAutowireTest {
 		assertEquals(expectedStart, externalUploadDestination.getUrl().substring(0, expectedStart.length()));
 	}
 	
+	@Test
+	public void testExternalS3UploadDestinationSecurityCheck() throws Exception {
+		String testBase = "test-base-" + UUID.randomUUID();
+		ExternalS3UploadDestinationLocationSetting externalS3LocationSetting = new ExternalS3UploadDestinationLocationSetting();
+		externalS3LocationSetting.setBanner("upload here");
+		externalS3LocationSetting.setDescription("external");
+		externalS3LocationSetting.setUploadType(UploadType.S3);
+		externalS3LocationSetting.setUrl(Constants.S3_HOSTNAME);
+		externalS3LocationSetting.setBucket(String.format(StackConstants.EXTERNAL_S3_TEST_BUCKET, StackConfiguration.getStack()));
+		externalS3LocationSetting.setBaseKey(testBase);
+
+		s3Client.createBucket(externalS3LocationSetting.getBucket());
+
+		try {
+			projectSettingsManager.createUploadDestinationLocationSetting(userInfo, externalS3LocationSetting);
+			fail();
+		} catch (IllegalArgumentException e) {
+			assertTrue(e.getMessage().indexOf("Did not find S3 object") != -1);
+		}
+
+		String nothing = "";
+		ObjectMetadata metadata = new ObjectMetadata();
+		metadata.setContentLength(nothing.length());
+		s3Client.putObject(externalS3LocationSetting.getBucket(), testBase + "owner.txt", new StringInputStream(nothing), metadata);
+		
+		S3FileHandle fauxHandle = new S3FileHandle();
+		fauxHandle.setBucketName(externalS3LocationSetting.getBucket());
+		fauxHandle.setKey(testBase + "owner.txt");
+		toDelete.add(fauxHandle);
+
+		try {
+			projectSettingsManager.createUploadDestinationLocationSetting(userInfo, externalS3LocationSetting);
+			fail();
+		} catch (IllegalArgumentException e) {
+			assertTrue(e.getMessage().indexOf("No username found") != -1);
+		}
+
+		String wrongName = "not me";
+		metadata.setContentLength(wrongName.length());
+		s3Client.putObject(externalS3LocationSetting.getBucket(), testBase + "owner.txt", new StringInputStream(wrongName), metadata);
+
+		try {
+			projectSettingsManager.createUploadDestinationLocationSetting(userInfo, externalS3LocationSetting);
+			fail();
+		} catch (IllegalArgumentException e) {
+			assertTrue(e.getMessage().indexOf("is not what was expected") != -1);
+		}
+
+		metadata.setContentLength(username.length());
+		s3Client.putObject(externalS3LocationSetting.getBucket(), testBase + "owner.txt", new StringInputStream(username), metadata);
+
+		externalS3LocationSetting = projectSettingsManager.createUploadDestinationLocationSetting(userInfo, externalS3LocationSetting);
+
+		// make sure only owner can use this s3 setting
+		UploadDestinationListSetting uploadDestinationListSetting = new UploadDestinationListSetting();
+		uploadDestinationListSetting.setProjectId(projectId);
+		uploadDestinationListSetting.setSettingsType(ProjectSettingsType.upload);
+		uploadDestinationListSetting.setLocations(Lists.newArrayList(externalS3LocationSetting.getUploadId()));
+
+		try {
+			projectSettingsManager.createProjectSetting(userInfo2, uploadDestinationListSetting);
+			fail();
+		} catch (UnauthorizedException e) {
+			assertTrue(e.getMessage().indexOf("Only the owner") != -1);
+		}
+
+		projectSettingsManager.createProjectSetting(userInfo, uploadDestinationListSetting);
+	}
+
 	@Test
 	public void testCreateCompressedFileFromString() throws Exception{
 		String fileContents = "This will be compressed";
@@ -575,4 +678,13 @@ public class FileHandleManagerImplAutowireTest {
 		return text;
 	}
 	
+	private void addAcl(String projectId, Long principalId) throws Exception {
+		AccessControlList acl = accessControlListDAO.get(projectId, ObjectType.ENTITY);
+		Set<ACCESS_TYPE> accessTypes = Sets.newHashSet(ACCESS_TYPE.READ, ACCESS_TYPE.UPDATE, ACCESS_TYPE.CREATE, ACCESS_TYPE.UPLOAD);
+		ResourceAccess ra = new ResourceAccess();
+		ra.setPrincipalId(principalId);
+		ra.setAccessType(accessTypes);
+		acl.getResourceAccess().add(ra);
+		accessControlListDAO.update(acl, ObjectType.ENTITY);
+	}
 }
